Extract handling of Japanese in SpellingVdalitor as JapaneseExpressionVariationValidator
