+++ redpen-core/src/main/java/cc/redpen/validator/sentence/SpellingValidator.java
@@ -21,7 +21,6 @@ import cc.redpen.model.Sentence;
 import cc.redpen.tokenizer.TokenElement;
 
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
@@ -31,29 +30,6 @@ public final class SpellingValidator extends SpellingDictionaryValidator {
 
     @Override
     public void validate(Sentence sentence) {
-        String lang = getSymbolTable().getLang();
-        if (lang.equals("ja")) {
-            validate_ja(sentence);
-        } else if (lang.equals("en")) {
-            validate_en(sentence);
-        }
-    }
-
-    private void validate_ja(Sentence sentence) {
-        for (TokenElement token : sentence.getTokens()) {
-            String reading = token.getTags().get(7);
-            if (this.words.containsKey(reading)) {
-                List<TokenElement> tokens = this.words.get(reading);
-                for (TokenElement candidate : tokens) {
-                    if (candidate != token && !token.getSurface().equals(candidate.getSurface())) {
-                        addLocalizedErrorFromToken(sentence, token);
-                    }
-                }
-            }
-        }
-    }
-
-    private void validate_en(Sentence sentence) {
         for (TokenElement token : sentence.getTokens()) {
             String surface = token.getSurface().toLowerCase();
             if (surface.length() == 0 || surface.matches("\\P{L}+")) continue;
@@ -61,23 +37,6 @@ public final class SpellingValidator extends SpellingDictionaryValidator {
             if (dictionaryExists() && !inDictionary(surface)) {
                 addLocalizedErrorFromToken(sentence, token);
             }
-        }
-    }
-
-    @Override
-    public void preValidate(Sentence sentence) {
-        String lang = getSymbolTable().getLang();
-        if (!lang.equals("ja")) {
-            return;
-        }
-
-        for (TokenElement token : sentence.getTokens()) {
-            String reading = token.getTags().get(7);
-            if (!this.words.containsKey(reading)) {
-                this.words.put(reading, new LinkedList<TokenElement>());
-            }
-            this.words.get(reading).add(token);
-
         }
     }
 }
