+++ redpen-core/src/test/java/cc/redpen/validator/sentence/JapaneseExpressionVariationValidatorTest.java
@@ -1,0 +1,52 @@
+/**
+ * redpen: a text inspection tool
+ * Copyright (c) 2014-2015 Recruit Technologies Co., Ltd. and contributors
+ * (see CONTRIBUTORS.md)
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package cc.redpen.validator.sentence;
+
+import cc.redpen.RedPen;
+import cc.redpen.RedPenException;
+import cc.redpen.config.Configuration;
+import cc.redpen.config.ValidatorConfiguration;
+import cc.redpen.model.Document;
+import cc.redpen.validator.BaseValidatorTest;
+import cc.redpen.validator.ValidationError;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+import java.util.Map;
+
+import static java.util.Collections.singletonList;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class JapaneseExpressionVariationValidatorTest extends BaseValidatorTest {
+    protected JapaneseExpressionVariationValidatorTest() {
+        super("JapaneseExpressionVariation");
+    }
+
+    @Test
+    void japanese() throws RedPenException {
+        config = Configuration.builder("ja")
+                .addValidatorConfig(new ValidatorConfiguration(validatorName))
+                .build();
+
+        Document document = prepareSimpleDocument("之は山です。これは川です。");
+
+        RedPen redPen = new RedPen(config);
+        Map<Document, List<ValidationError>> errors = redPen.validate(singletonList(document));
+        assertEquals(2, errors.get(document).size());
+    }
+}
